requirements.txt
Django>=4.0
graphene-django
django-crontab==0.7.1
gql[all]
